+++ src/app/loadout-builder/LoadoutBuilder.tsx
@@ -150,6 +150,7 @@ export default memo(function LoadoutBuilder({
     {
       loadoutParameters,
       statOrder,
+      autoStatMods,
       pinnedItems,
       excludedItems,
       subclass,
@@ -163,8 +164,6 @@ export default memo(function LoadoutBuilder({
   const isPhonePortrait = useIsPhonePortrait();
 
   const lockedExoticHash = loadoutParameters.exoticArmorHash;
-
-  const autoStatMods = loadoutParameters.autoStatMods ?? false;
 
   const lockedMods = useMemo(
     () =>